% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unconf_issues.R
\name{add_repo}
\alias{add_repo}
\title{Add repo to those tracked}
\usage{
add_repo(repos_df, repo_owner, repo_name, label)
}
\arguments{
\item{repo_owner}{}

\item{repo_name}{}

\item{label}{}
}
\description{
Add repo to those tracked
}
