% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unconf_issues.R
\name{get_labelled_issues}
\alias{get_labelled_issues}
\title{Get labelled issues from a specific repository}
\usage{
get_labelled_issues(repo_owner, repo_name, label, ...)
}
\arguments{
\item{repo_owner}{}

\item{repo_name}{}

\item{label}{}

\item{...}{}
}
\description{
Get labelled issues from a specific repository
}
