% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unconf_issues.R
\name{keep_issues_and_cols}
\alias{keep_issues_and_cols}
\title{Keep issues and columns}
\usage{
keep_issues_and_cols(df)
}
\arguments{
\item{df}{}
}
\description{
Only keep issues (no PRs) and relevant columns from pulled issues
}
