% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_issue_viewer.R
\name{mod_issue_viewer_ui}
\alias{mod_issue_viewer_ui}
\alias{mod_issue_viewer}
\title{mod_issue_viewer_ui and mod_issue_viewer_server}
\usage{
mod_issue_viewer_ui(id)

mod_issue_viewer(input, output, session, issue_type, collapsed = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
