% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unconf_issues.R
\name{get_repos_issues}
\alias{get_repos_issues}
\title{Get all unconf issues from specified repositories}
\usage{
get_repos_issues(repositories_path, issue_type)
}
\arguments{
\item{issue_type}{}

\item{repositories}{}
}
\description{
Get all unconf issues from specified repositories
}
